requirements.txt
torch
torchvision
torchaudio
matplotlib
numpy
opencv-python
Pillow
PySide6
jupyterlab
flask
Flask-RESTful
flask_sqlalchemy
pytest
pytest-cov
pytest-mock
reportlab
sqlalchemy
scikit-learn
scipy
seaborn

fastapi
uvicorn
torch
torchvision
opencv-python
numpy
pillow
python-multipart
